Adiciona componente ChipsContainer
